+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Added: `declaration-block-no-shorthand-property-overrides` rule.
 - Added: `declaration-block-trailing-semicolon` rule.
 - Added: `no-browser-hacks` rule.
+- Added: `no-unsupported-browser-features` rule.
 - Added: `selector-max-specificity` rule.
 - Added: `string-no-newline` rule.
 - Added: `named-where-possible` option to `font-weight-notation`.